BUG: Re-organize code and remove intermediate logic libraries
* Integrates the code from "QuickTCGALogic" and "ShortCutCoreLogic"
into the module logic. VTK wrapping of the non VTK class is disabled
setting the WRAP_EXCLUDE property.

* Rename "utilities.h" to "TCGAUtilities.h" to avoid clash with
other header.

* Split "TCGAUtilities" into header/cxx/hxx

* Associate include directories with "NucleiSegLib" using "Modern CMake"
command "target_include_directories" available with CMake >= 3.0. This
mean that the QuickTCGA logic linking against NucleiSegLib will automatically
be compiled with the appropriate set of include directories.

* Upon installation of "SlicerPathology" extension and its dependency
 "SlicerOpenCV", the module "QuickTCGA" and "ShortCutCore" along with
the associated editor effects are all loaded.
