+++ ShortCutCore/Logic/ShortCutCoreSegmenter.h
@@ -12,7 +12,7 @@
 #include "opencv2/highgui/highgui.hpp"
 #include "opencv2/imgproc/imgproc.hpp"
 
-#include "utilities.h"
+#include "TCGAUtilities.h"
 #include "ShortCut.h"
 
 class ShortCutCoreSegmenter {
